requirements.txt
asgiref==3.6.0
Django==3.2.17
django-jazzmin==2.6.0
django-js-asset==2.0.0
djangorestframework==3.14.0
psycopg2==2.9.5
pytz==2022.7.1
sqlparse==0.4.3
